Switch default 'edited' flag to True in PointService.py
The default status of the 'edited' flag for each 'point' data in PointService.py has been switched to True. This shift aims to optimize tracking of changes and handling of individual-point sequences by including this flag in determining 'seq_mark' and 'type'. On saving, the 'edited' flag is cleared for a clean data presentation.